create link
This change introduces the link type. Target, owner, and group
can all be managed (though owner and group are only effective on
some OSes). Target changes are done by deleting and recreating
the link, though this may be changed in the future to an atomic
move to avoid any race conditions.