create group type
This change adds the group type. Group additions without a specified
gid use the highest current gid + 1. The group password can be managed,
though this is highly discouraged by most distros. Group admins have not
yet been implemented. Members are currently not managed via the group type.

If a gid is equal to 0, libconman considers this value unset. On most
(possibly all?) distros, the root group is 0 and should probably not be
altered.