manage file contents
This change adds the ability to manage file contents. Two fields
were added to the cm_file struct, hash and source. Hash is populated
by file_get and is a SHA1 hash of the file contents. This is mostly
useful for verbose/noop runs and determining whether there is a
change. The source field is a file path string of the source file.

The ability to copy a file has, for now, been added directly to
set_state. This may be changed in the future but should not have
and affect on the API.