enable recursive owner, group, state, and mode changes
This change enables recursive management of directory contents. It
is now possible to manage owner, group, mode, and state recursively.
Currently, all recursive functions have the same signature and pass
around the path to apply the changes, the actual change arg, and a
pointer to itself. This could result in issues for large directory
hierarchies but some small tests did not encounter any major issues.