Enable package updating
This change enables the updating of packages by introducing the
CM_LATEST and CM_OUTDATED states. Additionally, a cm_diff_state utility
function has been added to assist with determining whether a change will
be made, so this does not need to be reimplemented every time.

cm_run_cmd has been modified to accept a NULL function pointer for cases
where we want to ignore the output of and exec and just return the exit
code.